Fixed toolbar problem
Description: When I scroll down, toolbar doesn't appear on top of the screen. That's because I didn't arrange the view properly. But I fixed this issue. Actually when we look at to the code, maybe it was unnecessary to build with ConcatAdapter, but if I need to add something besides rvProductListingScreen, I can do this without create another adapter. That's why I will keep this. Plus I haven't used this before. It's really cool.